@@ -6,7 +6,7 @@
     "ng": "ng",
     "serve": "node server",
     "start": "concurrently -c \"yellow.bold,green.bold\" -n \"SERVER,BUILD\" \"nodemon server\" \"ng build --watch\"",
-    "build": "ng build --prod",
+    "build:prod": "ng build --configuration production",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
@@ -14,16 +14,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.0.3",
+    "@angular/animations": "^12.0.0",
     "@angular/cdk": "^11.0.2",
-    "@angular/common": "^11.0.3",
-    "@angular/compiler": "^11.0.3",
-    "@angular/core": "^11.0.3",
-    "@angular/forms": "^11.0.3",
+    "@angular/common": "^12.0.0",
+    "@angular/compiler": "^12.0.0",
+    "@angular/core": "^12.0.0",
+    "@angular/forms": "^12.0.0",
     "@angular/material": "^11.0.2",
-    "@angular/platform-browser": "^11.0.3",
-    "@angular/platform-browser-dynamic": "^11.0.3",
-    "@angular/router": "^11.0.3",
+    "@angular/platform-browser": "^12.0.0",
+    "@angular/platform-browser-dynamic": "^12.0.0",
+    "@angular/router": "^12.0.0",
     "bcrypt": "^3.0.2",
     "body-parser": "^1.18.2",
     "compression": "^1.7.2",
@@ -47,29 +47,29 @@
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "rxjs": "^6.5.5",
+    "rxjs": "^7.0.1",
     "swagger-ui-express": "^3.0.9",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.3",
-    "@angular/cli": "^11.0.3",
-    "@angular/compiler-cli": "^11.0.3",
-    "@angular/language-service": "^11.0.3",
+    "@angular-devkit/build-angular": "~12.0.0",
+    "@angular/cli": "^12.0.0",
+    "@angular/compiler-cli": "^12.0.0",
+    "@angular/language-service": "^12.0.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~12.12.14",
+    "@types/node": "~15.0.0",
     "codelyzer": "^6.0.0",
     "concurrently": "^3.5.1",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ts-node": "~6.1.0",
     "tslint": "~6.1.0",
-    "typescript": "4.0.5"
+    "typescript": "4.2.4"
   }
 }
